Move `MusicDisplay` into `monument_cli`
